script: Move Reflector to script_bindings.
